requirements_dev.txt
pytest==7.1.3
python-dotenv==1.0.0
